- inheritance, import *, generator
